requirements.txt
streamlit
pandas
PyPDF2
openai
python-dotenv
Pillow
anthropic
requests
PyMuPDF